TINY-6284: Moved native fullscreen utilities into NativeFullscreen.ts
